feat(api): add support for oauth2 token endpoint
Cognito supports more features than just its AWS-API calls demonstrate.
This service fake simulates just those AWS APIs - except for
.well-known/jwks.json.

This PR starts to add support for the token endpoint outlined here:
https://docs.aws.amazon.com/cognito/latest/developerguide/token-endpoint.html

However, since we are not serving static data, we need to create a fake
target, and then add this to our AWS API Router, so that we can get
access to the services objects.